gee: use PR description as merge commit message, improve PR template. (#347)
gee: use PR description as merge commit message, improve PR template.

As a consequence of Enfabrica's "Better Code Reviews" discussion, we now have a
better idea of what should be covered in any PR description.  We've also
affirmatively decided that the first PR comment is the PR description, which
should also be used as the merge-commit message when the PR is submitted.

In this PR, "gee pr_make"'s behavior is changed to mandate the production of
a PR description that fills in a specified template.  We check that description
to make sure that it at least contains a Title line and a first paragraph.

Testing: I'm using the new gee script to create and later submit this PR.

See also:
  "Enfabrica Code Review Culture" aka "Better Code Reviews"
  https://docs.google.com/document/d/1sfooxl0GcaAqgl79_52pLscSVwJw5RWFUvgacRUyRwI/edit?usp=sharing

PR generated by jonathan from branch gee_pr_description.