gee codeowners: fix for "WORKSPACES" (tiny PR) (#444)
gee codeowners: fix for "WORKSPACES" (tiny PR)

gee codeowners: fix for "WORKSPACES" (tiny PR)

I noticed that `gee codeowners` wasn't reporting the right codeowners for
changes to WORKSPACES.  Turns out this was an edge-condition in my rules
processing code that only affects rules that aren't wildcards, and the
file in question happens to be in the root directory of the repo.

While I'm in here, I also snuck in a purely cosmetic improvement to
the markdown format of the generated codeowners comment.

Tested:

- made a test branch and only changed the WORKSPACES file
- verified that "gee codeowners" reported the right owners.