add version check to gee version (#560)
gee version: tell user if a new version is available.

gee version: tell user if a new version is available.

The `gee version` command will now check for a newer version of gee on astore,
and tell the user how to upgrade if a newer version is available.

Tested: Manually ran ./gee version to view message.